Move ReplacedBy to org.gradle.api
Make sure we get an error if you combine two annotations with ReplacedBy
