+++ subprojects/core-api/src/main/java/org/gradle/api/ReplacedBy.java
@@ -14,7 +14,9 @@
  * limitations under the License.
  */
 
-package org.gradle.api.tasks;
+package org.gradle.api;
+
+import org.gradle.api.tasks.Internal;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -25,13 +27,13 @@ import java.lang.annotation.Target;
 /**
  * <p>Attached to a task property to indicate that the property has been replaced by another. Like {@link Internal}, the property is ignored during up-to-date checks.</p>
  *
- * <p>This annotation should be attached to the getter <b>and</b> setter method in Java or the property field in Groovy. You should also consider adding {@link Deprecated} to any replaced property.</p>
+ * <p>This annotation should be attached to the getter method in Java or the property field in Groovy. You should also consider adding {@link Deprecated} to any replaced property.</p>
  *
  * <p>This will cause the task <em>not</em> to be considered out-of-date when the property has changed.</p>
  */
 @Documented
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
+@Target({ElementType.METHOD, ElementType.FIELD})
 public @interface ReplacedBy {
     /**
      * The Java Bean-style name of the replacement property.
